bio_info/replication/protein-phenotype/pyproject.toml
[project]
name = "protein-phenotype"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dx-client",
    "datalake",
    "plotnine>=0.15.3",
    "polars>=1.39.3",
    "pyarrow>=23.0.1",
    "pyiceberg>=0.11.1",
    "pyspark>=4.1.1",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "statsmodels>=0.14.6",
    "torch>=2.11.0",
    "tqdm>=4.67.3"
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
